requirements.txt
absl-py==1.0.0
astor==0.8.1
attrs==21.4.0
cached-property==1.5.2
certifi==2021.10.8
click==8.0.3
cycler==0.11.0
efficientnet==0.0.4
Flask==2.0.2
fonttools==4.29.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.43.0
h5py==2.10.0
imageio==2.15.0
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mediapipe==0.8.9.1
networkx==2.6.3
numpy==1.21.5
opencv-contrib-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.2.0
PyYAML==6.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
typing_extensions==4.0.1
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
